buffes/zio-buf-kmalloc: the default buffer (only input by now)
This is the default buffer: it manages a list of data blocks
(the lenght is fixed at 16 blocks by now) for any device and
any trigger.  The actual file operations in this case are
generic, and live in zio-buf.c (zio-core.ko).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Signed-off by: Federico Vaga <federico.vaga@gmail.com>