general: input and output is always cset-based
This changes some method and stuff, in order to make all input and
output be cset-based. This solves a number of issues with the previous
setup. Thus, even drivers must be made differentlym abd this fixes
zio-zero accordingly.

As a side effect, we have a data_done method in trigger ops, so the device
can tell when it is done. And the fire_trigger returns void, since it
cannot wait for stuff to be done. See documentation for details.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>

docs for general-cset