zio.h: add a few fields to cset and channel
The cset needs a spinlock for I/O operations. Locking the whole device
seems overdue.

The cset and channel need some private data, because they are
allocated as arrays (while ti and bi don't need it, as container_of
can be used).  This adds a device-private pointer to both cset
and channel, and a trigger-private pointer to the channel (this is
used e.g. by the ldisc demultiplexer, to be committed later).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>