core: ti and bi use the device-level spinlock
ZIO uses the device-level spinlock to protect device registers
during access; a buffer or a trigger instance can be assigned to
only one device on ZIO and if they are hardware implemented they
also live on the same perihperals. So we serialize all the
configuration operations on device, trigger and buffer by using
zio_device->lock.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>