core: offer generic_fops, add v_op
We will have buffers that use the default operations (the majority) and
buffers without file operations for in-kernel activities, like the
input subsystem. In the latter case we want the cdev to refuse being
openend, so NULL is reserved for that case.

Also, add v_op to the buffer, so it can support mmap.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>