bugfix: assign bi and ti after their registration
The bug was evident in zio_change_current_buffer. When you change the
buffer, zio registers a new bi, but the old __bi_register() overwrites
the current bi in use by the channel; so, when zio tries to unregister
the old bi, there is not any reference to it.

Now you must assign a bi to a channel after its registration. For
simmetry and to prevent future bugs, also ti assignment must be done
after __ti_registration().

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>