bugfix: clear ZTI_BUSY flag when trigger is disabled
When disabling a trigger, ZIO invokes t_op->abort() to stop acquisition.
If the abort function invokes data_done() to return partial
blocks, the ZTI_BUSY flag is correctly cleared. But, if the abort
function frees all active blocks the flag remains set.

With this patch, always clear the ZTI_BUSY flag when the trigger is
disabled so it is correct if/when the trigger is re-enabled.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>