core/locks: serialize trig->create and zbuf->create; init instance lock
The core should ensure its users that create calls are not reentrant.
We use the locks that already exist in the buffer type and the trigger
type.

Then, we init both ti->lock and bi->lock, so our users can rely on
them (see next patch).

Related to locks: fix a comment in zio.h

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>