core: improve abort and data_done mechanism
The helper zio_trigger_abort() is introduced; it can be used to abort a
data transfer.

The bit flag ZTI_COMPLETING is introduced: if high ZIO is completing the
transfert operation by invokng the data_done() function; if low the
transfer is running or the trigger is waiting for event. If abort() is
called when this glag is high, the functions does nothing.

The helper zio_trigger_data_done() is introduced; insted of call the
data_done() function with cset->ti->t_op->data_done() and manually set
the ZTI_COMPLETING flag, this hellper must be used. If the trigger
operation data_done is not set, then use the __zio_internal_data_done

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>