incompatible change: define addr_zio and use it in the control
This adds 4 bytes to the control (there are atill 16 free) to fit
a whole struct addr_zio (same as sockaddr_zio, to be defined when we
merge PF_ZIO). The structure was almost ready for that,
but we missed the sa_family 16-bit field.

This also changes the minor number for the ZIO control structure, and
all the few users of the address-related fields.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>