core: don't module_get() for internal buffer/trigger
If a driver module registerscreate its own preferred trigger or
buffer, we can't increment its own use count, or we won't be able to
unload the module.

This patch resolves the issue. Now the ZIO core doesn't increment the
use counter of the trigger or buffer, if it's owned by the device.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>