sysfs: replace nsamples with pre-samples and post-samples
It is difficult to support a board which has pre-samples and
post-samples registers and synchronize these value with the nsamples
zio_attribute and control field. But, it is really simple for board that
hasn't pre and post samples to use these two attributes because they can
use only one attribute for their purpose (only pre-sample or only
post-sample).

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>