buffers/kmalloc: fix buffer-full check, as size may change
If the user changes the buffer size, this might make more space for a
sleeping writer. Similarly, the buffer may be fuller than the new
value after a change, so we check for ">= limit" and not "== limit".

This doesn't apply to the vmalloc buffer, that currently doesn't allow
a size change.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>