tools/zio-dump: be grep-friendly and not over-verbose
Changes attribute print-out to only show attributes that are
defined (using the mask).  Show them as both decimal and hex.

All lines are grep-friendly, as the used to be. You can grep for
"Ctrl:" or "Data:" or specific attributes.

Example (with "-a"):

   Ctrl: version 0.7, trigger user, dev zloop-0000, cset 1, chan 1
   Ctrl: seq 18, n 16, size 1, bits 8, flags 01000001 (little-endian)
   Ctrl: stamp 1351773544.171337188 (0)
   Ctrl: trigger-std-mask: 0x0002
   Ctrl: trigger-std-1   0x00000010        16
   Data: 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00

For those who, sometimes, want to see all attributes, even when the
mask is zero, there is the "-A" option.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>