zio-sysfs.h: change zio_attribute structure (incompatible)
This patch removes the union field priv in the zio_attribute and it
introduces the id field. From our experience, the priv.ptr pointer
is useless; the priv.addr field can be used for every purpose, also
if you have your own private structure.

The new field id replaces the role of priv.addr. It can represent a
register address, a register offset, an index of a vector of private
structure which describe how to gain access to a register, or whatever
can describe uniquely a register (or register field)

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>