core: write ctrl->nsamples every time the input trigger fires
The driver itself is allowed to change nsamples, because it may have
less samples than requested when the trigger fires. This is done by
zio-mini, for example, that will be merged soon.  Therefore, the
number in current_ctrl must be updated, or the previous shorter value
will stick.

Also, move seq_num incrementing earlier, so even if allocation fails
the sequence number is updated (and we need to do alarms too for that).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>