trigger: new trigger operation arm to handle trigger arming
This commit adds a new trigger operation: arm.  The operation is
useful because it allows a trigger to define its own procedure to arm
itself. It is used by the fmc-adc-100Ms-14bit-4cha driver.

At the moment, nobody use the trigger operation data_done() and abort()
because the generic one work in most of the case, and tipically there
is nothing special to do if the arming procedure is always the same.
This patch make sense of these operations because if you need a custom
arm(), probably, you also need a custom data_done() and abort().

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>