core: incompatible: buffer->alloc_block must allocate the control
This is an incompatible change for buffer authors. We don't think
there are any out-of-tree buffers, though.  The allocation
function must allocate the control by itself, instead of
receiving it from outside.  The reason why the control was passed
doesn't hold any more, and the upcoming PF_ZIO buffer will
greatly benefit from this.

Also, now alloc_block returns NULL on error, so it can be used directly
by the caller (it used to be ERR_PTR).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>