doc/img: more figures (also logo), and change formats
pdftex has the very bad habit of picking png first, so we can't
have png images or the pdf version will be horrible. Thus, let's use
gif for the html (which is not picked by pdftex) and a .pdf generated
by our encapsulated postscript sources.

This adds the needed makefile and removes the previous pdf images,
which are regenerated in the next commit. We are also adding new
images where, including the logo used in the cover.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>