buffers/zio-buf-vmalloc.c: add a "merge-data" attribute, document it
The merge-data attribute tells the buffer to try a merge of a
newly-stored block with the previous one. When this is possible, the
buffer updates the previous block (nsamples, datalen) and releases the
new control immediately. This allows saving some memory without
breaking the ZIO data model. The feature only makes sense for input,
but there is no policy in place to enforce that.

When user space reads this new block, it will get attributes (and
timestamp) for the first data items, but all available items are
returned in this bigger block.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>