core: accept zero-size time csets
This commit allows zero-size channels to exist. The feature is used
for TDC and DTC (time to digital converters and the opposite).  This
simple change in the core allows to implement TDC drivers that only
return the control (which includes a time stamp) and no data.

For DTC to work we need some modification to chardev.c, which are not
there yet.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>