triggers/zio-trig-irq: two minor fixes
These changes allow the driver to work with a PCI interrupt, a role
that was never used earlier. The commit fixes a warning on removal
when if libgpio is not present in the host kernel; it also tries to
request the shared interrrupt with various edge-trigger bits, until
one succeeds. This allows to use both edges for owned GPIO interrupts
and fall back to no specific requirement when used for a PCI
interrupt.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>