bugfix for buffer change (long standing race condition)
When the current buffer is changed, there should be no
users on the channels (otherwise, either a user block
is being produced/consumed or there's somebody sleeping
on the wq of the buffer instance).

This prevents changing the buffer if in use, and prevents
users from opening a channel if the buffer is being changed.
We use the "disabled" flag for the buffer, not used so far.

We noticed this while working on buffer changes (see previous
commit). I'm unable to backport this to the proper place in the
history because it relies on the new locking introduced by this branch.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>