trigger timer: add phase attribute
The ms-phase attribute pairs with ms-period. In this way you can have
a fine control about when the I/O happens.  Please note that the epoch
of the time base is unphatomable (it is related to boot time). So you
can manage offsets between csets, but not absolute stamps. For that
please use the HRT trigger.

Every trigger instance starts at phase 0 (i.e., the trigger is armed
at a multiple of its period). If you later change the period, no phase
calculation is performed, so that the new period will be as requested
(to allow for slowly-changing periods.  If, however, you change the
phase, the trigger is re-armed with that absolute phase, to allow for
slowly-changing phase.  If you change the period and then you change
the phase, you'll experience an unexpected period during the transition.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>