sysfs: add 'devtype' sysfs attribute to all zio object
The Linux kernel automatic updates the uevent attributes only if the
device has a bus or a class. When the Linux kernel updates the uevent it
automatically adds the DEVTYPE attribute if the device has a
device_type. In ZIO cset, channel, buffer and trigger have not a bus or
a class, but they have a different device_type. These attributes are used
to export the device_type to user-space and allow libudev to easy
inspect the ZIO device hierarchy.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>