triggers: soft timestamp is in  zio_arm_trigger()
Soft timestamping happens at zio_arm_trigger time. This removes
the various software timestamping in the individual trigger.
Moreover, arming is idempotent, so timestamp in the locked section,
only when the trigger is actually armed.

As usual, hardware can overwrite ti timestamps, so this software stamp
is only used for devices that have no better notion of time.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>