helpers.c: on zio_trigger_data_done, call stop_io to invorm device
When the acquisition is over, the driver invokes zio_trigger_data_done()
to store blocks and make some clean up on trigger. But it can be useful
for the driver to clean up itself, or it must force the hardware to
interrupt the I/O.

So, with this patch, ZIO use the stop_io() operation both: when the trigger
aborts and when an acquisition is over. This should be not a problem for
the driver because in both cases the data transfer must be stopped.

More over, simmetry with the function  zio_arm_trigger(), which invoks
cset->raw_io(), the function zio_trigger_data_done() must invoke
cset->stop_io().

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>