sysfs: new std attribute 'version' for device, trigger and buffer
The 'version' attribute is a standard attribute. It is placed at
the last position of the standard attribute array. In this way, it is
easier to identify it in an hexdump stream because it is the last
standard attribute. At the moment the last index of standard attributes
array is 15, but in a far away future may change.

The aim of this attribute is to identify the version of an attribute set
of a ZIO object (device, trigger, buffer). During development the set
of attribute exported by a driver may change. This attribute must be
used to notify these changes to user-space programs.
The attribute is optional.

When used in a device, we suggest to place it at device level.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>