[bug-fix] bus.c: fix device probe failure
The bug: if a device probe() fails, ZIO register the real ZIO device
anyway
The issue: on device probe() channel-set, trigger, channels and buffer
are not ready

The patch: do probe after the _complete_ registration of the real
device. In this way, it is guaranteed that all components are ready.
More over, if the probe fail, the device can be easly removed.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>