buffer/sock-nic.c: use trigger helper to push block
This patch use zio_trigger_try_push() to push a block to the trigger
instead of custom code. More over, it remove the error handling
condition because it is not necessary to free block on store error, it
is already done internally.

There is no way to know if the store fail, so we cannot increment
tx_dropped counter.