bugfix in buffers: block might be used after being freed
[This commit originated in zio-v1.0-fixed, and I then adapted to
the current master]

With trigger user and output channels, a block may be referenced after
being released. This affects both kmalloc and vmalloc buffers.

This bug was introduced by commit 2300e624, partly reversed by this.
However, we only noticed when running under CONFIG_DEBUG_SLAB.

Bug description:

  When store_block is called for output the first time, it pushes to
  the trigger, but as of commit 2300e624 it adds the block to the
  list first (to keep ordering if another push happens very soon).
  If push succeeds *and* cset->raw_io succeeds immediately, the
  block is freed.  Store_block will thus remove it from the list
  of enqueued blocks, accessing the data structure.

Bug solution:

  The store_block method runs calls trigger->push in locked context,
  like it used to before 2300e624. Thus, the block doesn't need
  to be enqueued (as other store processes are waiting for the
  spinlock) and is not dequeued if push succeeds.

The ZIO_BI_PUSHING flag remains, to prevent the deadlock when
store succeeds and data_done() tries to retrieve the next block.
However, it is checked by retr_block before taking the lock.

Update for master branch: the commit also removes the locking in
the new helpers zio_buffer_retr_block() and zio_buffer_store_block(),
because the functions are now called in locking context; suck a lock
is not needed anyways, because the critical section is just reading the
DISABLED flag, which is atomic by definition.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>