zio-buf-vmalloc: allow size changes, when not mapped
The commits allows the vmalloc buffer to change the allocated size.
If mmap is active, it returns EBUSY, but we need to protect
against new maps to appear during the change. The vm_ops->open
method returns void, so we need to block the mmap() system call
directly, in chardev.c.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>