zio: add current len and kb sysfs buf attribute
This patch add two new attributes: "allocated-buffer-len" and
"allocated-buffer-kb". Buffers are expected to use one of them, or both.

The purpose of these attribute is to export to user space information about
the number of blocks stored in a buffer. This information can be used to
monitor buffer status, and see how the data flow behaves.

We report the allocated buffer size, which includes active blocks,
instead of the size currently stored, because it reflects actual use
of the available data space (the max-buffer-* attributes). Knowing how
many blocks, or kilobytes, are currently stored is much less
important.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>