zio-trig-hrt.c: change relative time from 1h to 10s
The in-kenrel hrt uses absolute times, and so does the zio trigger
based on it.  However, for testing in shell scripts, relative times
are easier to handle.  We accepted any "absolute" time less than 1 hour
to be considered relative, but then we realized some embedded systems
have no real-world time notion, so they start at Jan 1st 1970.

This decreases relative times to 10s, assuming it's enough for shell
scripts during testing and short enough to allow the rare embedded system
to be later than tat by the time zio runs.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>