bugfix: general: use GFP_ATOMIC for registration-time allocs
Registration of devices is done with a mutex held. as device_attach()
calls device_lock() (in the upstream kernel).

Thus, allocation of minor numbers must use GFP_ATOMIC. Similarly, the
default buffer and the default trigger for a new device being
registered are allocated within the critical section.

The problem also applies when changing a buffer or trigger type
at runtime, because sysfs_write_file() takes a mutex on the file.
Thus, all triggers and buffers are affected.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>