core [bugfix]: on output fail, remove block
The trigger-user fires _only_ on ti->push(), and buffers try to
push blocks _only_ if there are no enqueued blocks.

When using the trigger-user on the output flow and the correspondent
cset->raw_io() operation fails, then the pushing does not succed and
buffers start to enqueue the incoming blocks instead of pushing them
to the trigger.

This patch frees all active_blocks if cset->raw_io() fails (like we
did for the input flow), and it raise the trigger lost alarm.

The patch also raise the trigger lost alarm on the input flow and
clears the active_block

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>
Acked-by: Alessandro Rubini <rubini@gnudd.com>