buffer: disable trigger while resize vmalloc
If an arm occurs while we are resizing the buffer it may gets the wrong
pointers.

For example, this is the case when we use the 'user' trigger. Because we
want to remove all blocks from the buffer before resize it, we are
calling retr_block() which internally does pull() and the trigger user
on pull calls arm().

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>