headers: add new string length
ZIO_OBJ_NAME_LEN is *just* the limit for names, but sometimes we want to append
thing to a given name. Typically, index to identify the component in the
hierarchy. So this patch introduce ZIO_OBJ_NAME_FULL_LEN which should be used as
size for any string that modify the original name

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>