sysfs: fix symbolic link creation
The introduction on symbolic link was not 100% correct on patch

af1c35a  Enumerate channels and channel-sets in sysfs

If a custom name is not assigned, DO NOT force one (like we did). Leave it
empty so that we can easily recognize this condition.

Change the sysfs 'name' attribute in order to use the proper variable

Signed-off-by: federico.vaga@cern.ch <federico.vaga@gmail.com>